+++ sdk/handler.go
@@ -21,12 +21,14 @@ type Handler struct {
 }
 
 // NewHandler creates a new Handler with an http mux.
+// It configures a single default route for the plugin activation and adds a bunch of middlewares.
 func NewHandler(logger logrus.FieldLogger, manifest string) Handler {
 	mux := handlers.NewRouter(logger)
 	mux.Use(handlers.ErrorMiddleware)
+	mux.Use(errorMiddleware)
+	mux.Use(contentTypeMiddleware)
 
 	mux.HandleFunc(activatePath, func(w http.ResponseWriter, r *http.Request, p map[string]string) error {
-		w.Header().Set("Content-Type", DefaultContentTypeV1_1)
 		fmt.Fprintln(w, manifest)
 		return nil
 	})
@@ -90,6 +92,11 @@ func (h Handler) ServeWindows(addr, pluginName, daemonDir string, pipeConfig *Wi
 		defer os.Remove(spec)
 	}
 	return h.Serve(l)
+}
+
+// Use adds a middleware at the beginning of the middleware stack. The lastly added middleware is called first.
+func (h Handler) Use(m handlers.Middleware) {
+	h.mux.Use(m)
 }
 
 // HandleFunc registers a function to handle a request path with.